% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mare_slopes.R
\name{calc_mare_slopes}
\alias{calc_mare_slopes}
\title{Calculate median absolute relative error and slopes
Function to calculate MARE and slopes for simulated values}
\usage{
calc_mare_slopes(input)
}
\arguments{
\item{input}{Input, should have cpue of zero at depletion of zero and nhooks column}
}
\description{
Calculate median absolute relative error and slopes
Function to calculate MARE and slopes for simulated values
}
