% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_two.R
\name{change_two}
\alias{change_two}
\title{Function to modify two factors
Function to modify two factors}
\usage{
change_two(thing1, thing2, name1, name2, ctl, ncores = 6, add_index = FALSE)
}
\arguments{
\item{thing1}{Vector of things to loop over}

\item{thing2}{Vector of things to loop over}

\item{name1}{Name of thing1 in the ctl file}

\item{name2}{Name of thing1 in the ctl file}

\item{ncores}{Number of cores to use in computation}

\item{ctl_in}{Control file to modify}
}
\description{
Function to modify two factors
Function to modify two factors
}
