% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_locs.R
\name{define_locs}
\alias{define_locs}
\title{Define Fishing Locations
Function to define fishing locations based on certain conditions}
\usage{
define_locs(numrow = 10, numcol = 10, condition)
}
\arguments{
\item{numrow}{Number of rows in fishing area}

\item{numcol}{Number of columns in fishing area}

\item{condition}{Condition used to subset the data. Input as text then parsed out
Subsets based on the expanded grid}
}
\description{
Define Fishing Locations
Function to define fishing locations based on certain conditions
}
\examples{
define_locs(condition = "y == 3 & x == 5")
}
