% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_movement.R
\name{define_movement}
\alias{define_movement}
\title{Define Movement range of each location
Function to define movement range at specified locations}
\usage{
define_movement(fish_area, x, y, scope)
}
\arguments{
\item{fish_area}{Area of fish}

\item{x}{X location of fishing}

\item{y}{Y location of fishing}

\item{scope}{Scope of fish movement}
}
\description{
Define Movement range of each location
Function to define movement range at specified locations
}
