% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filled.contour2.R
\name{filled.contour2}
\alias{filled.contour2}
\title{Filled Contour 2
Modification of filled.contour function}
\usage{
filled.contour2(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y,
  finite = TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim,
  nlevels), nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, mar, ...)
}
\arguments{
\item{there}{are paremeters}
}
\description{
Filled Contour 2
Modification of filled.contour function
}
