% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pop_loop.R
\name{fish_pop_loop}
\alias{fish_pop_loop}
\title{Fish population in a loop function
Function to fish population in a for loop.}
\usage{
fish_pop_loop(fish_area, loc_row, ctl = ctl)
}
\arguments{
\item{fish_area}{List that has the matrices of fish, needs to be list}

\item{loc_row}{Locations of fishing}

\item{ctl}{ctl file, from make_ctl function}
}
\description{
Fish population in a loop function
Function to fish population in a for loop.
}

