% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_parallel.R
\name{fixed_parallel}
\alias{fixed_parallel}
\title{Function that takes fixed inputs and runs scenarios
Function that calls run_scenario. Allows for parallelization across greater number of cores}
\usage{
fixed_parallel(index, ctl1, to_loop = to_loop, change_these)
}
\arguments{
\item{index}{Index of to_loop to subset}

\item{ctl1}{Starting ctl list}
}
\description{
Function that takes fixed inputs and runs scenarios
Function that calls run_scenario. Allows for parallelization across greater number of cores
}
