% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_plot_input.R
\name{format_plot_input}
\alias{format_plot_input}
\title{Format Plot Input
Function to format the simulation output for plots.}
\usage{
format_plot_input(out)
}
\arguments{
\item{out}{output from conduct_survey function}
}
\description{
Format Plot Input
Function to format the simulation output for plots.
}
\examples{
ctl <- make_ctl(distribute = 'uniform', mortality = .1, move_out_prob = .5,
                nfish1 = 1000, nfish2 = 10000, prob1 = 1, prob2 = .3, nyear = 15, scope = 1)
out <- conduct_survey(ctl = ctl) 
format_plot_input(out)
}
