% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rand_locs.R
\name{get_rand_locs}
\alias{get_rand_locs}
\title{Get Random Fishing Locations}
\usage{
get_rand_locs(seed = 3, numrow = 10, numcol = 10, nlocs = 20)
}
\arguments{
\item{seed}{Set seed}

\item{numrow}{Number of rows in fishing area}

\item{numcol}{Number of columns in fishing area}

\item{nlocs}{Maximum number of fishing locations}
}
\description{
Function to generate random fishing locations.
}
\examples{
rand_locs <- get_rand_locs()
}
