% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_population.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize Population}
\usage{
initialize_population(ctl, nfish)
}
\arguments{
\item{ctl}{List of control parameters from make_ctl function, description of arguments in 
make_ctl function}

\item{nfish}{Number of fish, use this to generate matrices for both species}
}
\description{
Initialize the spatial distribution of the fish population
}
\examples{
Uniformly distribute fish
control <- make_ctl()
initialize_population(ctl = control)
Distribute fish in upper right quadrant only
control <- make_ctl(distribute = 'area', area = 'upperright')
initialize_population(ctl = control)

Patchily distribute fish
control <- make_ctl(distribute = 'patchy')
initialize_population(ctl = control)

}

