% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_back.R
\name{move_back}
\alias{move_back}
\title{Move fish back
Function that moves fish back to their starting place. This occurs after fishing}
\usage{
move_back(nfish_moved, samps_out, kk = 0)
}
\arguments{
\item{nfish_moved}{List of fish that moved from previous function}

\item{samps_out}{Fish samples from fishing step}
}
\description{
Move fish back
Function that moves fish back to their starting place. This occurs after fishing
}

