% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish.R
\name{move_fish}
\alias{move_fish}
\title{Move Fish
Function to move fish}
\usage{
move_fish(fish_range1 = fish_range, nfish_outside1 = nfish_outside,
  zero_index1 = zero_index)
}
\arguments{
\item{fish_range1}{The range of fish to calculate the numbers of fish from. Based on the "scope"
parameter specified in the ctl file}

\item{nfish_outside1}{The number of fish outside the range}

\item{zero_index1}{The location of fishing activity}
}
\description{
Move Fish
Function to move fish
}
