% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_sites.R
\name{pick_sites}
\alias{pick_sites}
\title{Pick Sites
Function to pick number of good, medium, and bad sites}
\usage{
pick_sites(nbest = 0, nmed = 0, nbad = 0, fish_mat)
}
\arguments{
\item{nbest}{Number of best/good sites to samle}

\item{nmed}{Number of medium sites to samle}

\item{ctl}{Control file}

\item{nbest}{Number of bad sites to samle}
}
\description{
Pick Sites
Function to pick number of good, medium, and bad sites
}
