% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_sites_prob.R
\name{pick_sites_prob}
\alias{pick_sites_prob}
\title{Pick Sites based on probability
Function to pick number of good, medium, and bad sites}
\usage{
pick_sites_prob(nsites, fish_mat, samp_option)
}
\arguments{
\item{nsites}{Number of sites to select}

\item{fish_mat}{Matrix of fish}

\item{samp_option}{Sample option, either random or for best locations}
}
\description{
Pick Sites based on probability
Function to pick number of good, medium, and bad sites
}
