% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_cpue.R
\name{plot_average_cpue}
\alias{plot_average_cpue}
\title{Ploat Average CPUE
Function to plot the average cpue. Calls format_plot_input function}
\usage{
plot_average_cpue(out)
}
\description{
Ploat Average CPUE
Function to plot the average cpue. Calls format_plot_input function
}
\examples{
ctl <- make_ctl(distribute = 'uniform', mortality = .1, move_out_prob = .5,
                nfish1 = 1000, nfish2 = 10000, prob1 = 1, prob2 = .3, nyear = 15, scope = 1)
out <- conduct_survey(ctl = ctl) 
plot_average_cpue(out)
}
