% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_replicates.R
\name{run_replicates}
\alias{run_replicates}
\title{Run replicates of things
Function to run replicates of change_two. Only save things for plot. This is because other issues
should have been addressed before running multiple iterations}
\usage{
run_replicates(niters, thing1, name1, thing2, name2, ncores,
  add_index = FALSE, ctl)
}
\arguments{
\item{niters}{Number of iterations}

\item{thing1}{Vector of things to loop over}

\item{name1}{Name of thing1 in the ctl file}

\item{thing2}{Vector of things to loop over}

\item{name2}{Name of thing1 in the ctl file}

\item{ncores}{Number of cores to use in computation}

\item{add_index}{Option to add Index of things to loop over, applicaple when looping over numbers of locations}

\item{ctl}{Control file to modify}
}
\description{
Run replicates of things
Function to run replicates of change_two. Only save things for plot. This is because other issues
should have been addressed before running multiple iterations
}
