% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampled_locs.R
\name{run_sampled_locs}
\alias{run_sampled_locs}
\title{Run Simulation with preferential and random location sampling
Run Simulation with preferential and random location sampling}
\usage{
run_sampled_locs(thing1, name1, nreps, nsites_vec, shape_list, ncores, ctl_o)
}
\arguments{
\item{thing1}{thing1 vector}

\item{name1}{name of thing1 to loop over}

\item{nreps}{Number of replicates}

\item{nsites_vec}{Vecotr of the number of sites to sample}

\item{shape_list}{List of shapes for beta distribution}

\item{ncores}{Number of cores}

\item{ctl_o}{Input ctl list}
}
\description{
Run Simulation with preferential and random location sampling
Run Simulation with preferential and random location sampling
}
