% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampled_locs_2spp.R
\name{run_sampled_locs_2spp}
\alias{run_sampled_locs_2spp}
\title{Run Simulation with preferential and random location sampling for two species and sampled locations
Run Simulation with preferential and random location sampling for two species and sampled locations}
\usage{
run_sampled_locs_2spp(thing1, thing2, name1, name2, nreps, nsites_vec,
  shape_list, ncores, ctl_o)
}
\arguments{
\item{thing1}{thing1 vector}

\item{thing2}{thing2 vector}

\item{name1}{name of thing1 to loop over}

\item{name2}{name of thing2 to loop over}

\item{nreps}{Number of replicates}

\item{nsites_vec}{Vecotr of the number of sites to sample}

\item{shape_list}{List of shapes for beta distribution}

\item{ncores}{Number of cores}

\item{ctl_o}{Input ctl list}
}
\description{
Run Simulation with preferential and random location sampling for two species and sampled locations
Run Simulation with preferential and random location sampling for two species and sampled locations
}
