% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run Scenarios
Function to run scenarios in parallel. Returns three things, a list of all the output from
conduct_survey, cpue summarized by each value in loop_over, and summarized data for each year.}
\usage{
run_scenario(ctl_start, loop_over, ncores = 1, to_change, add_index = FALSE,
  par_func)
}
\arguments{
\item{ctl_start}{Output from make_ctl function}

\item{loop_over}{Vector of values to loop over}

\item{ncores}{Number of cores to run in parallel, default is 1}

\item{to_change}{Specify which value to modify in lapply statement. Should be the values in}

\item{add_index}{Specify add index or not, for locations}

\item{par_func}{Turn parallel run option on or off
loop_over and be a character string}
}
\description{
Run Scenarios
Function to run scenarios in parallel. Returns three things, a list of all the output from
conduct_survey, cpue summarized by each value in loop_over, and summarized data for each year.
}
\examples{
ctl1 <- make_ctl(distribute = 'patchy', mortality = 0, move_out_prob = .5,
      nfish1 = 0, nfish2 = 0, prob1 = .01, prob2 = 0, nyear = 15, scope = 1, seed = 4,
      location = one_loc, numrow = 1, numcol = 1)  
ttest <- run_scenario(ctl = ctl, loop_over = seq(100, 1500, by = 100), to_change = 'nfish1', 
  ncores = 6)
}
