% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_change.R
\name{sample_change}
\alias{sample_change}
\title{sample_change
Function to calculate power of the survey}
\usage{
sample_change(nsamps = 1000, dep_fixed, dep_vec, input)
}
\arguments{
\item{nsamps}{Number of samples}

\item{dep_fixed}{Fixed depletion level to compare power}

\item{dep_vec}{Vector of depletion levels to compare against dep_vec}

\item{input}{Input data that contains simulation results}
}
\description{
sample_change
Function to calculate power of the survey
}
