% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_exp.R
\name{sample_exp}
\alias{sample_exp}
\title{Exponential Fish Sampling
Function that samples based on exponential function. Catch probabilities depend on number
of fish and probabilities of catching each species. 
Function to fish the population}
\usage{
sample_exp(nfish1, nfish2, prob1, prob2)
}
\arguments{
\item{nfish1}{Number of fish1}

\item{nfish2}{Number of fish2}

\item{prob1}{Probability of catching fish1}

\item{prob2}{Probability of catching fish2}
}
\description{
Exponential Fish Sampling
Function that samples based on exponential function. Catch probabilities depend on number
of fish and probabilities of catching each species. 
Function to fish the population
}
\examples{
Put Example Here
}

