% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_good_locs.R
\name{sample_good_locs}
\alias{sample_good_locs}
\title{Sample Good Locations
Conduct survey in a proportion of locations with fish.}
\usage{
sample_good_locs(ctl, prop_good, ngoods, which_spp)
}
\arguments{
\item{ctl}{Control file}

\item{prop_good}{Proportion of good site to sample}

\item{ngoods}{Number of total good sites}

\item{which_spp}{Specify spp1 or spp2}
}
\description{
Sample Good Locations
Conduct survey in a proportion of locations with fish.
}
