% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send email that run is done
Function to send an email, arguments won't match}
\usage{
send_email(from = "<pkuriyama@gmail.com>", to = "<pkuriyama@gmail.com>",
  subject = "run done", body = "run done", mailcontrol = list(smtpServer =
  "ASPMX.L.GOOGLE.COM"))
}
\arguments{
\item{ctl}{Control file}

\item{nbest}{Number of best/good sites to samle}

\item{nmed}{Number of medium sites to samle}

\item{nbest}{Number of bad sites to samle}
}
\description{
Send email that run is done
Function to send an email, arguments won't match
}
