% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_fishing.R
\name{visualize_fishing}
\alias{visualize_fishing}
\title{Visualize Fishing location
Function to visualize fishing location}
\usage{
visualize_fishing(loc, fish_mat)
}
\arguments{
\item{loc}{Fishing location}

\item{fish_mat}{Matrix of initial fish distribution}
}
\description{
Visualize Fishing location
Function to visualize fishing location
}
