% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cpue.R
\name{calc_cpue}
\alias{calc_cpue}
\title{Calculate CPUE
Function to calculate cpue}
\usage{
calc_cpue(survey_res, ctl)
}
\arguments{
\item{survey_res}{Output from conduct_survey function}

\item{cpue_method}{Method of calculating CPUE. Options are 'average' for calculating
straight average by year.}
}
\description{
Calculate CPUE
Function to calculate cpue
}

