% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locs.R
\name{run_locs}
\alias{run_locs}
\title{Run Different Seeds with location scenarios
Function takes number of sites of differing quality and repeats replicate runs for different
seeds}
\usage{
run_locs(shape_list, loc_scenario, loc_list, loc_vector, ncores, ctl_o, thing1,
  name1, par_func = "change_two", index1 = FALSE, index2 = TRUE)
}
\arguments{
\item{shape_list}{Shape Scenarios to run, must specify}

\item{loc_scenario}{Specify "pick" or "increasing"}

\item{loc_list}{List of locations to loop 
through, if loc_scenario == 'pick'}

\item{loc_vector}{Vector of number of locations to sample, 
if loc_scenario == 'increasing'}

\item{ncores}{Number of cores}

\item{ctl_o}{Original ctl list}

\item{thing1}{Thing1 to loop over, see change_two function}

\item{name1}{Thing1 to loop over, see change_two function}
}
\description{
Run Different Seeds with location scenarios
Function takes number of sites of differing quality and repeats replicate runs for different
seeds
}
