% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locs.R
\name{run_locs}
\alias{run_locs}
\title{Run Different Seeds with location scenarios
Function takes number of sites of differing quality and repeats replicate runs for different
seeds}
\usage{
run_locs(nbests = 5, nmeds = 5, nbads = 0, seeds, ncores, nsites = 10,
  ctl_o, thing1, name1, par_func = "change_two")
}
\arguments{
\item{nbests}{Number of best sites, pick_sites function}

\item{nmeds}{Number of medium sites, pick_sites function}

\item{nbads}{Number of bad sites, pick_sites function}

\item{seeds}{Number of seeds}

\item{ncores}{Number of cores}

\item{nsites}{Number of total sites}

\item{ctl_o}{Original ctl list}

\item{thing1}{Thing1 to loop over, see change_two function}

\item{name1}{Thing1 to loop over, see change_two function}
}
\description{
Run Different Seeds with location scenarios
Function takes number of sites of differing quality and repeats replicate runs for different
seeds
}
