% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locs_2spp.R
\name{run_locs_2spp}
\alias{run_locs_2spp}
\title{Loop over locations with two species
Function to loop over locations with two species}
\usage{
run_locs_2spp(shape_list, loc_scenario, loc_list, loc_vector, ncores, ctl_o,
  thing1, name1, thing2, name2, par_func = "change_two", index1 = FALSE,
  index2 = FALSE)
}
\arguments{
\item{shape_list}{Shape Scenarios to run, must specify}

\item{loc_scenario}{Specify "pick" or "increasing"}

\item{loc_list}{List of locations to loop 
through, if loc_scenario == 'pick'}

\item{loc_vector}{Vector of number of locations to sample, 
if loc_scenario == 'increasing'}

\item{ncores}{Number of cores}

\item{ctl_o}{Original ctl list}

\item{thing1}{Thing1 to loop over, see change_two function}

\item{name1}{Name of thing1 to loop over, see change_two function}

\item{thing2}{Thing2 to loop over, see change_two function}

\item{name2}{Name of thing2 to loop over, see change_two function}
}
\description{
Loop over locations with two species
Function to loop over locations with two species
}
